requirements.txt
fastapi==0.115.11
sqlalchemy==2.0.38
greenlet==3.1.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
aiofiles==24.1.0
uuid==1.30
asyncio==3.4.3
uvicorn[standart]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20

loguru==0.7.3
nano==0.10.0

ruff==0.9.9
isort==6.0.1
flake8==7.1.2
pylint==3.3.5

pytest==8.3.5
pytest-mock==3.14.0
pytest-asyncio==0.25.3
httpx==0.28.1
